changed some datatypes to float

+++ backend/db/migrations/20220829211209-create-spot.js
@@ -26,10 +26,10 @@ module.exports = {
         type: Sequelize.STRING(255)
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       name: {
         type: Sequelize.STRING(255)
@@ -38,7 +38,8 @@ module.exports = {
         type: Sequelize.STRING(255)
       },
       price: {
-        type: Sequelize.INTEGER
+        // price can include decimals
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
